+++ examples/json_assert.php
@@ -1,0 +1,11 @@
+<?php
+
+use SamMcDonald\Jason\Assert\JsonAsserter;
+
+require_once '../vendor/autoload.php';
+
+
+JsonAsserter::assertStringIsValidJson('{"foo":}');
+
+echo "\n";
+echo "\n";