+++ examples/json_assert_2.php
@@ -1,0 +1,11 @@
+<?php
+
+use SamMcDonald\Jason\Assert\JsonAsserter;
+
+require_once '../vendor/autoload.php';
+
+
+JsonAsserter::throwInvalidArgument('foo');
+
+echo "\n";
+echo "\n";