+++ examples/load_from_file.php
@@ -1,15 +1,10 @@
 <?php
 
+require_once '../vendor/autoload.php';
 
 use SamMcDonald\Jason\Jason;
 
-require_once '../vendor/autoload.php';
-require_once './SampleClasses/User.php';
-require_once './SampleClasses/History.php';
-
-
 echo Jason::fromFile('./JsonExamples/profile.json');
-
 
 echo "\n";
 echo "\n";